+++ pytestqt/wait_signal.py
@@ -168,7 +168,7 @@ class SignalBlocker(_AbstractSignalBlocker):
 
         .. note:: contrary to the parameter of same name in
             :meth:`pytestqt.qtbot.QtBot.waitSignal`, this parameter does not
-            consider the :ref:`qt_wait_signal_raising`.
+            consider the :ref:`qt_default_raising` option.
 
     :ivar list args:
         The arguments which were emitted by the signal, or None if the signal
@@ -608,12 +608,114 @@ class SignalEmittedSpy(object):
                 )
 
 
+class CallbackBlocker(object):
+
+    """
+    .. versionadded:: 3.1
+
+    An object which checks if the returned callback gets called.
+
+    Intended to be used as a context manager.
+
+    :ivar int timeout: maximum time to wait for the callback to be called.
+
+    :ivar bool raising:
+        If :class:`TimeoutError` should be raised if a timeout occured.
+
+        .. note:: contrary to the parameter of same name in
+            :meth:`pytestqt.qtbot.QtBot.waitCallback`, this parameter does not
+            consider the :ref:`qt_default_raising` option.
+
+    :ivar list args:
+        The arguments with which the callback was called, or None if the
+        callback wasn't called at all.
+
+    :ivar dict kwargs:
+        The keyword arguments with which the callback was called, or None if
+        the callback wasn't called at all.
+    """
+
+    def __init__(self, timeout=1000, raising=True):
+        self.timeout = timeout
+        self.raising = raising
+        self.args = None
+        self.kwargs = None
+        self.called = False
+        self._loop = qt_api.QtCore.QEventLoop()
+        if timeout is None:
+            self._timer = None
+        else:
+            self._timer = qt_api.QtCore.QTimer(self._loop)
+            self._timer.setSingleShot(True)
+            self._timer.setInterval(timeout)
+
+    def wait(self):
+        """
+        Waits until either the returned callback is called or timeout is
+        reached.
+        """
+        __tracebackhide__ = True
+        if self.called:
+            return
+        if self._timer is not None:
+            self._timer.timeout.connect(self._quit_loop_by_timeout)
+            self._timer.start()
+        self._loop.exec_()
+        if not self.called and self.raising:
+            raise TimeoutError("Callback wasn't called after %sms." % self.timeout)
+
+    def _quit_loop_by_timeout(self):
+        try:
+            self._cleanup()
+        finally:
+            self._loop.quit()
+
+    def _cleanup(self):
+        if self._timer is not None:
+            _silent_disconnect(self._timer.timeout, self._quit_loop_by_timeout)
+            self._timer.stop()
+            self._timer = None
+
+    def __call__(self, *args, **kwargs):
+        # Not inside the try: block, as if self.called is True, we did quit the
+        # loop already.
+        if self.called:
+            raise CallbackCalledTwiceError("Callback called twice")
+        try:
+            self.args = list(args)
+            self.kwargs = kwargs
+            self.called = True
+            self._cleanup()
+        finally:
+            self._loop.quit()
+
+    def __enter__(self):
+        return self
+
+    def __exit__(self, type, value, traceback):
+        __tracebackhide__ = True
+        if value is None:
+            # only wait if no exception happened inside the "with" block
+            self.wait()
+
+
 class SignalEmittedError(Exception):
     """
     .. versionadded:: 1.11
 
     The exception thrown by :meth:`pytestqt.qtbot.QtBot.assertNotEmitted` if a
     signal was emitted unexpectedly.
+    """
+
+    pass
+
+
+class CallbackCalledTwiceError(Exception):
+    """
+    .. versionadded:: 3.1
+
+    The exception thrown by :meth:`pytestqt.qtbot.QtBot.waitCallback` if a
+    callback was called twice.
     """
 
     pass